Realtime Segment Relocation Manager (#2604)
Introducing a service, which will periodically scan ideal state of all realtime tables, and move segments which are completed, from the consuming servers to the completed servers. The tags to use for relocation will be provided by the tenant config (soon to be introduced), which will specify the realtime consuming tag and the realtime completed tag.